% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource.R
\name{all_datasources}
\alias{all_datasources}
\title{Tibble of all datasources connected to workbook}
\usage{
all_datasources(twb_file)
}
\arguments{
\item{twb_file}{a Tableau twb file}
}
\value{
a tibble of catasources connected to workbook
}
\description{
Tibble of all datasources connected to workbook
}
\examples{
 \dontrun{
   all_datasources(twb_file = "test.xml")
}

}
