% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{all_raw_fields}
\alias{all_raw_fields}
\title{Create an overview of all visible fields from a twb file}
\usage{
all_raw_fields(twb_file)
}
\arguments{
\item{twb_file}{an XML file generated from a twb}
}
\value{
Returns a tibble with all fields coming from the datasource(s) used in the workbook.
}
\description{
Create an overview of all visible fields from a twb file
}
\examples{
 \dontrun{
   all_raw_fields(twb_file = "test.xml")
}
}
