% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{all_windows}
\alias{all_windows}
\title{Tibble of all dashboards, worksheets and storyboards}
\usage{
all_windows(twb_file)
}
\arguments{
\item{twb_file}{a Tableau TWB file}
}
\value{
a tibble with a list of dashboards, worksheets and storyboards
}
\description{
Tibble of all dashboards, worksheets and storyboards
}
\examples{
 \dontrun{
   other_created_overview(twb_file = "test.xml")
}

}
