% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_name}
\alias{check_name}
\title{Checks the "caption" created to ensure no duplicates}
\usage{
check_name(twb_file, a_name)
}
\arguments{
\item{twb_file}{a twb file}

\item{a_name}{a name for a calculation}
}
\value{
returns TRUE/FALSE depending on if the name exists in the dataset
}
\description{
Checks the "caption" created to ensure no duplicates
}
\examples{
 \dontrun{
   check_name(twb_file, a_name)
}
}
