% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_calc.R
\name{create_calc}
\alias{create_calc}
\title{Add a calculation to a Tableau workbook}
\usage{
create_calc(
  twb_file,
  file_location,
  caption,
  datatype,
  role,
  type,
  class = "Tableau",
  formula
)
}
\arguments{
\item{twb_file}{A tableau file}

\item{file_location}{location of where the twb file is stored}

\item{caption}{user-friendly name}

\item{datatype}{string, number, date, ...}

\item{role}{measure or dimension}

\item{type}{nominal, ordinal or ...}

\item{class}{calculations are a type tableau}

\item{formula}{formula for the calculation}
}
\value{
an updated twb_file with an added calculation
}
\description{
Add a calculation to a Tableau workbook
}
\examples{
 \dontrun{
   create_calc(twb_file, file_Location, caption, datatype,
   role, type, class, formula)
}
}
