% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_calc_id}
\alias{create_calc_id}
\title{create unique ID for the calculation.}
\usage{
create_calc_id(twb_file)
}
\arguments{
\item{twb_file}{twb_file}
}
\value{
a unique code for a calculation
}
\description{
create unique ID for the calculation.
}
\examples{
 \dontrun{
   create_calc_id(twb_file)
}
}
\keyword{internal}
