% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calc_template.R
\name{create_calc_template}
\alias{create_calc_template}
\title{Creates a template to enter all calculations for the workbook}
\usage{
create_calc_template(csv_file)
}
\arguments{
\item{csv_file}{the location and file name for the template}
}
\value{
an empty csv file with all options for creating a calculation
}
\description{
Creates a template to enter all calculations for the workbook
}
\examples{
 \dontrun{
   create_calc_template(csv_file)
}
}
