% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{create_toc}
\alias{create_toc}
\title{Title Create a table of Contents based on the workbook}
\usage{
create_toc(twb_file)
}
\arguments{
\item{twb_file}{an xml file}
}
\value{
Table of contents
}
\description{
Title Create a table of Contents based on the workbook
}
\examples{
  \dontrun{
   create_toc(twb_file = "test.xml")
}
}
