% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_metadata.R
\name{extract_all_metadata}
\alias{extract_all_metadata}
\title{Function that extracts all metadata available from functions as a list}
\usage{
extract_all_metadata(twb_file)
}
\arguments{
\item{twb_file}{A TWB file}
}
\value{
a list of metadata extracted using the TWB2R functions
}
\description{
Function that extracts all metadata available from functions as a list
}
\examples{
  \dontrun{
   extract_all_metadata(twb_file = "test.twb")
}
}
