% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_all_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Returns the desired metadata based on the result}
\usage{
get_metadata(result_list, result_type)
}
\arguments{
\item{result_list}{All metadata.  Use extract_all_metadata for the result_list}

\item{result_type}{select from windows, raw, parameters, created (calculations, groups and bins),
datasources}
}
\value{
meta-data based on a keyword
}
\description{
Returns the desired metadata based on the result
}
\examples{
  \dontrun{
   get_metadata(twb_file = "test.twb")
}
}
