% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{other_created_overview}
\alias{other_created_overview}
\title{List of other fields created in tableau}
\usage{
other_created_overview(twb_file)
}
\arguments{
\item{twb_file}{a tableau twb file}
}
\value{
list of other created files
}
\description{
List of other fields created in tableau
}
\examples{
 \dontrun{
   other_created_overview(twb_file = "test.xml")
}
}
