% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{parameters_overview}
\alias{parameters_overview}
\title{Create overview of parameters}
\usage{
parameters_overview(twb_file)
}
\arguments{
\item{twb_file}{a twb file from tableau}
}
\value{
list of parameters
}
\description{
Create overview of parameters
}
\examples{
 \dontrun{
   data_dictionary(twb_file = "test.xml")
}
}
