% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_datasources.R
\name{show_all_datasources}
\alias{show_all_datasources}
\title{Tibble of all datasources connected to workbook}
\usage{
show_all_datasources(twb_file)
}
\arguments{
\item{twb_file}{a TWB file}
}
\value{
a tibble of datasources connected to workbook
}
\description{
Tibble of all datasources connected to workbook
}
\examples{
 \dontrun{
   show_all_datasources(twb_file = "test.twb")
}

}
