% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{show_all_other_created}
\alias{show_all_other_created}
\title{List of other fields created in tableau}
\usage{
show_all_other_created(twb_file)
}
\arguments{
\item{twb_file}{a tableau twb file}
}
\value{
list of other created files
}
\description{
List of other fields created in tableau
}
\examples{
 \dontrun{
   all_other_created(twb_file = "test.xml")
}
}
