% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{show_all_parameters}
\alias{show_all_parameters}
\title{Creates an overview of parameters created in the workbook}
\usage{
show_all_parameters(twb_file)
}
\arguments{
\item{twb_file}{a TWB file}
}
\value{
Returns a tibble of all parameters created in the workbook.
}
\description{
Creates an overview of parameters created in the workbook
}
\examples{
 \dontrun{
   show_all_parameters(twb_file = "test.twb")
}
}
