% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{show_all_raw_fields}
\alias{show_all_raw_fields}
\title{Create an overview of all visible fields from a TWB file}
\usage{
show_all_raw_fields(twb_file)
}
\arguments{
\item{twb_file}{a TWB file}
}
\value{
Returns a tibble of all fields that exist in datasource(s) used in the workbook.
}
\description{
Create an overview of all visible fields from a TWB file
}
\examples{
 \dontrun{
   all_raw_fields(twb_file = "test.twb")
}
}
