% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_windows.R
\name{show_all_windows}
\alias{show_all_windows}
\title{Tibble of all dashboards, worksheets and storyboards}
\usage{
show_all_windows(twb_file)
}
\arguments{
\item{twb_file}{a WB file}
}
\value{
a tibble with a list of dashboards, worksheets and storyboards
}
\description{
Tibble of all dashboards, worksheets and storyboards
}
\examples{
 \dontrun{
   show_all_windows(twb_file = "test.twb")
}

}
