% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{create_dd}
\alias{create_dd}
\title{Create a list consisting of raw fields and calculated fields}
\usage{
create_dd(twb_file)
}
\arguments{
\item{twb_file}{A twb file}
}
\value{
A list of raw and calculated fields from the tableau workbook
}
\description{
Create a list consisting of raw fields and calculated fields
}
\examples{
 \dontrun{
   create_dd(twb_file = "test.xml")
}
}
