% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\title{Create Data Dicationary from column and calculation tags}
\usage{
data_dictionary(twb_file)
}
\arguments{
\item{twb_file}{An xml file}
}
\value{
returns a data dictionary
}
\description{
Create Data Dicationary from column and calculation tags
}
