% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{show_all_parameters}
\alias{show_all_parameters}
\title{Create overview of parameters}
\usage{
show_all_parameters(twb_file)
}
\arguments{
\item{twb_file}{a twb file from tableau}
}
\value{
list of parameters
}
\description{
Create overview of parameters
}
\examples{
 \dontrun{
   all_parameters(twb_file = "test.xml")
}
}
